+++ src/SConscript
@@ -247,6 +247,8 @@ wesnoth_sources = Split("""
     attack_prediction.cpp
     build_info.cpp
     carryover.cpp
+    chat_command_handler.cpp
+    chat_events.cpp
     commandline_options.cpp
     config_cache.cpp
     controller_base.cpp
