+++ src/game_initialization/multiplayer_ui.hpp
@@ -104,6 +104,7 @@ public:
 	void set_location(const SDL_Rect& rect);
 	using widget::set_location;
 	const std::vector<std::string>& user_list() const { return user_list_; }
+	void send_to_server(const config& cfg) override;
 protected:
 	int xscale(int x) const;
 	int yscale(int y) const;
